发布商品部分功能完善

+++ src/com/bigoranger/bigoranger/PublishGoodsActivity.java
@@ -17,6 +17,7 @@ import org.json.JSONObject;
 
 import com.bigoranger.bigoranger.adapter.AddressAdapter;
 import com.bigoranger.bigoranger.adapter.CategoryAdapter;
+import com.bigoranger.bigoranger.adapter.ImageAdapter;
 import com.bigoranger.bigoranger.domain.Address;
 import com.bigoranger.bigoranger.domain.Category;
 import com.bigoranger.bigoranger.domain.Service;
@@ -66,7 +67,7 @@ import android.widget.Toast;
 public class PublishGoodsActivity extends Activity implements OnClickListener {
 	private static final int TAKE_PHOTO = 1;
 	private static final int SELECT_IMAGE = 2;
-	
+
 	private static final String IMAGE = "BigOranger" + File.separator + "Image";
 	private String picPath = ""; // 图片路径
 	public static String name; // 文件名称
@@ -75,9 +76,9 @@ public class PublishGoodsActivity extends Activity implements OnClickListener {
 	private String fileName = ""; // 文件绝对路径
 	private File file; // 文件
 	private Uri u; // 从相册获取图片的uri
-	
-	private int imageCount = 0;	//图片数量
-	
+
+	private int imageCount = 0; // 图片数量
+
 	private static DisplayMetrics metrics; // 手机屏幕分辨率
 	ContentResolver cr;
 
@@ -85,7 +86,7 @@ public class PublishGoodsActivity extends Activity implements OnClickListener {
 	private int index;
 	HashMap<Integer, String> chkMap = new HashMap<Integer, String>();
 	private com.bigoranger.bigoranger.MyGridView myGirdView; // 网格
-	
+
 	private Button takePhoto, selectImage, saveBtn, backBtn;
 	private EditText titleEdit, priceEdit, costPriceEdit, presentationEdit;
 	private Spinner category_Spinner, address_Spinner, tradeWay_Spinner;
@@ -97,7 +98,7 @@ public class PublishGoodsActivity extends Activity implements OnClickListener {
 		super.onCreate(savedInstanceState);
 		super.setContentView(R.layout.uploadpic);
 		init();
-		
+
 		selectImage.setOnClickListener(this);
 		takePhoto.setOnClickListener(this);
 		myGirdView.setOnItemClickListener(new OnItemClickListenerImpl());
@@ -119,8 +120,7 @@ public class PublishGoodsActivity extends Activity implements OnClickListener {
 
 	/**
 	 * 
-	 * Function：用户点击添加商品页面初始化
-	 * author：lionel
+	 * Function：用户点击添加商品页面初始化 author：lionel
 	 */
 	@SuppressLint("HandlerLeak")
 	private void init() {
@@ -140,19 +140,18 @@ public class PublishGoodsActivity extends Activity implements OnClickListener {
 		tradeWay_Spinner = (Spinner) findViewById(R.id.tradeWay);
 		myGirdView = (MyGridView) findViewById(R.id.myGridView);
 		service_Linearlayout = (LinearLayout) findViewById(R.id.service_Linerlayout);
-		
+
 		String[] tItems = getResources().getStringArray(R.array.TradeWay);
 		ArrayAdapter<String> tradAdapter = new ArrayAdapter<String>(this,
 				android.R.layout.simple_spinner_item, tItems);
 		tradeWay_Spinner.setAdapter(tradAdapter);
-		
-		
+
 		final Handler handler = new Handler() {
 			@Override
 			public void handleMessage(Message msg) {
 				String jsonString = (String) msg.obj;
 				JsonForResult result = JsonUtil.parseJsonFirst(jsonString);
-				
+
 				List<Category> categories = result.getCategories();
 				List<Address> addresses = result.getAddresses();
 				List<Service> services = result.getServices();
@@ -164,38 +163,41 @@ public class PublishGoodsActivity extends Activity implements OnClickListener {
 
 				category_Spinner.setAdapter(cateAdapter);
 				address_Spinner.setAdapter(addrAdapter);
-				
+
 				Iterator<Service> iterator = services.iterator();
-				
-				while(iterator.hasNext()){
+
+				while (iterator.hasNext()) {
 					Service service = iterator.next();
 					CheckBox chk = new CheckBox(PublishGoodsActivity.this);
 					chk.setId(Integer.parseInt(service.getId()));
 					chk.setText(service.getTitle());
 					service_Linearlayout.addView(chk);
 					chk.setOnCheckedChangeListener(new OnCheckedChangeListener() {
-						
+
 						@Override
-						public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-							if(isChecked){
-								chkMap.put(buttonView.getId(), (String) buttonView.getText());
+						public void onCheckedChanged(CompoundButton buttonView,
+								boolean isChecked) {
+							if (isChecked) {
+								chkMap.put(buttonView.getId(),
+										(String) buttonView.getText());
 							}
 						}
-					});	
+					});
 				}
 			}
 		};
-		
-		Thread thread = new Thread(new Runnable() {
-			@Override
-			public void run() {
-				String jsonString = HttpUtil.sendJson2Server(null, URLs.ADD_GOODS_URL);
-				Message msg = Message.obtain();
-				msg.obj = jsonString;
-				handler.sendMessage(msg);
-			}
-		});
-		thread.start();
+
+		// Thread thread = new Thread(new Runnable() {
+		// @Override
+		// public void run() {
+		// String jsonString = HttpUtil.sendJson2Server(null,
+		// URLs.ADD_GOODS_URL);
+		// Message msg = Message.obtain();
+		// msg.obj = jsonString;
+		// handler.sendMessage(msg);
+		// }
+		// });
+		// thread.start();
 
 	}
 
@@ -221,27 +223,26 @@ public class PublishGoodsActivity extends Activity implements OnClickListener {
 		}
 
 	}
-	
+
 	/**
 	 * 
-	 * Function：添加商品保存信息
-	 * author：lionel
+	 * Function：添加商品保存信息 author：lionel
 	 */
-	private void saveGoods(){
-		
+	private void saveGoods() {
+
 		String title = titleEdit.getText().toString().trim();
 		String price = priceEdit.getText().toString().trim();
 		String costPrice = costPriceEdit.getText().toString().trim();
 		String presentation = presentationEdit.getText().toString().trim();
 		int categoryId = category_Spinner.getSelectedView().getId();
 		int addressId = address_Spinner.getSelectedView().getId();
-		int tradeWayId = (int) tradeWay_Spinner.getSelectedItemId()+1;
+		int tradeWayId = (int) tradeWay_Spinner.getSelectedItemId() + 1;
 		StringBuilder sb = new StringBuilder();
 		Set<Integer> chkIds = chkMap.keySet();
 		Iterator<Integer> iterator = chkIds.iterator();
-		while(iterator.hasNext()){
+		while (iterator.hasNext()) {
 			sb.append(iterator.next());
-			if(iterator.hasNext()){
+			if (iterator.hasNext()) {
 				sb.append("|");
 			}
 		}
@@ -259,34 +260,36 @@ public class PublishGoodsActivity extends Activity implements OnClickListener {
 			saveJson.put("AddressId", addressId);
 			saveJson.put("TradeWay", tradeWayId);
 			saveJson.put("Server", serviceId);
-			
-			final Handler saveHandler = new Handler(){
+
+			final Handler saveHandler = new Handler() {
 				@Override
 				public void handleMessage(Message msg) {
 					String jsonString = (String) msg.obj;
 					String result = JsonUtil.parseJsonAddGoods(jsonString);
-					Toast.makeText(PublishGoodsActivity.this, result, Toast.LENGTH_SHORT).show();
+					Toast.makeText(PublishGoodsActivity.this, result,
+							Toast.LENGTH_SHORT).show();
 				}
 			};
-			
-			new Thread(new Runnable(){
+
+			new Thread(new Runnable() {
 
 				@Override
 				public void run() {
-					String jsonString = HttpUtil.sendJson2Server(saveJson, URLs.SAVE_GOODS_INFO_URL);
+					String jsonString = HttpUtil.sendJson2Server(saveJson,
+							URLs.SAVE_GOODS_INFO_URL);
 					Message msg = Message.obtain();
 					msg.obj = jsonString;
 					saveHandler.sendMessage(msg);
 				}
-				
+
 			}).start();
-			
+
 		} catch (JSONException e) {
 			e.printStackTrace();
 		}
-		
+
 	}
-	
+
 	/**
 	 * 
 	 * Function：图片和普通上传 author：lionel
@@ -295,9 +298,10 @@ public class PublishGoodsActivity extends Activity implements OnClickListener {
 		if ((picPath != null && picPath.length() > 0)) {
 			String userid = userId;
 			int goodsid = goodsId;
-			String param = "?userid="+userid+"&goodsid="+goodsid;
-			UploadFileTask uploadFileTask = new UploadFileTask(this, URLs.IMAGE_UPLOAD_URL + param);
-			Log.e("http",URLs.IMAGE_UPLOAD_URL + param);
+			String param = "?userid=" + userid + "&goodsid=" + goodsid;
+			UploadFileTask uploadFileTask = new UploadFileTask(this,
+					URLs.IMAGE_UPLOAD_URL + param);
+			Log.e("http", URLs.IMAGE_UPLOAD_URL + param);
 			uploadFileTask.execute(picPath);
 		}
 	}
@@ -323,7 +327,7 @@ public class PublishGoodsActivity extends Activity implements OnClickListener {
 
 			intent.putExtra(MediaStore.EXTRA_OUTPUT, uri);
 			picPath = file.getPath();
-			fileName=picPath;
+			fileName = picPath;
 			startActivityForResult(intent, TAKE_PHOTO);
 		} else {
 			new AlertDialog.Builder(PublishGoodsActivity.this)
@@ -354,24 +358,24 @@ public class PublishGoodsActivity extends Activity implements OnClickListener {
 	/**
 	 * 回调执行的方法，拍照或是选择照片后的自定义操作...
 	 */
+	@SuppressLint("ShowToast")
 	@Override
 	protected void onActivityResult(int requestCode, int resultCode, Intent data) {
 		if (requestCode == TAKE_PHOTO) {
 			if (resultCode == Activity.RESULT_OK) {
 				String str = null;
-	             if(!"".equals(fileName)){
-	            	 str = fileName.substring(
-	 						fileName.lastIndexOf("/") + 1, fileName.lastIndexOf("."));
-	             }
-				String newName = str+"mini";				
+				if (!"".equals(fileName)) {
+					str = fileName.substring(fileName.lastIndexOf("/") + 1,
+							fileName.lastIndexOf("."));
+				}
+				String newName = str + "mini";
 				try {
 					// 以字节流的形式压缩，以便上传到服务器！！
 					saveCompressPic(fileName, newName);
 					initImageAdapter(newName);// 初始化图片适配器
 					picPath = fileName;
-					Log.e("picPath",picPath);
 				} catch (Exception e) {
-					e.printStackTrace();
+					throw new RuntimeException("图片压缩失败！" + e);
 				}
 			}
 		}
@@ -390,7 +394,7 @@ public class PublishGoodsActivity extends Activity implements OnClickListener {
 								.getColumnIndexOrThrow(MediaStore.Images.Media.DATA);
 						cursor.moveToFirst();
 						String path = cursor.getString(colunm_index);
-						
+
 						if (path.endsWith("jpg") || path.endsWith("png")
 								|| path.endsWith("jpeg")) {
 							picPath = path;
@@ -399,7 +403,9 @@ public class PublishGoodsActivity extends Activity implements OnClickListener {
 									.toString();
 							String newName = str + "_mini";
 							// 以字节流的形式压缩，以便上传到服务器！！
-							saveCompressPic("", newName);
+							// saveCompressPic(str, newName);
+							saveCompressPic(picPath, newName);
+							Log.e("newName", newName);
 							initImageAdapter(newName);// 初始化图片适配器
 							String toPath = sdcardRoot + IMAGE + File.separator;
 							copyFile(picPath, toPath, str + ".jpg");// 将选择的图片复制到指定路径下
@@ -412,13 +418,16 @@ public class PublishGoodsActivity extends Activity implements OnClickListener {
 					}
 
 				} catch (Exception e) {
-					e.printStackTrace();
+					// throw new RuntimeException("相册选择图片失败！请重新尝试..."+e);
+					Toast.makeText(PublishGoodsActivity.this,
+							"你选择的图片不存在，请重新选择", Toast.LENGTH_SHORT);
+					selectImage();
 				}
 			}
 		}
 		super.onActivityResult(requestCode, resultCode, data);
 		imageCount++;
-		picUpload(JsonUtil.userid,JsonUtil.goodsid);
+		// picUpload(JsonUtil.userid,JsonUtil.goodsid);
 	}
 
 	/**
@@ -427,12 +436,12 @@ public class PublishGoodsActivity extends Activity implements OnClickListener {
 	 * @param fname
 	 */
 	private void initImageAdapter(String fname) {
-
+		// 传入文件名，添加文件路径到list
 		list.add(sdcardRoot + IMAGE + File.separator + fname + ".jpg");
 		if (list.size() > 0 && list.size() < 9) {
 			myGirdView.setAdapter(new ImageAdapter(PublishGoodsActivity.this,
 					list, metrics.widthPixels));
-			// 显示上传进度
+			// 2秒后显示上传进度
 			showProgressDialog();
 		} else {
 			Toast.makeText(getApplicationContext(), "最多只能上传8张图片",
@@ -502,14 +511,14 @@ public class PublishGoodsActivity extends Activity implements OnClickListener {
 	 * @throws Exception
 	 */
 	private void saveCompressPic(String path, String newName) throws Exception {
-		
+
 		File file_02 = new File(sdcardRoot + IMAGE + File.separator, newName
 				+ ".jpg");
 		FileOutputStream out = new FileOutputStream(file_02);
 
 		Bitmap oldBitmap;
 		// 调用方法，获取压缩图片
-		if (path!=null||!"".equals(path)) {
+		if (path != null || !"".equals(path)) {
 			oldBitmap = BitmapFactory.decodeFile(path);// 调用拍照得到的图片
 			// 获取图片的旋转角度，有些系统把拍照的图片旋转了，有的没有旋转
 			int degree = Camera.readPictureDegree(path);
@@ -584,25 +593,27 @@ public class PublishGoodsActivity extends Activity implements OnClickListener {
 					String text = titleEdit.getText().toString().trim();
 
 					if (text != null && text.length() > 0) {
-						Map<String,String> map = new HashMap<String, String>();
+						Map<String, String> map = new HashMap<String, String>();
 						map.put("Title", text);
 						final JSONObject jsonObject = new JSONObject(map);
-						final Handler getCate_handler = new Handler(){
+						final Handler getCate_handler = new Handler() {
 							@Override
 							public void handleMessage(Message msg) {
 								String jsonString = (String) msg.obj;
-								List<Category> categories = JsonUtil.parseJsonGetCategory(jsonString);
+								List<Category> categories = JsonUtil
+										.parseJsonGetCategory(jsonString);
 								category_Spinner.setAdapter(null);
-								CategoryAdapter new_CateAdapter = new CategoryAdapter(categories, PublishGoodsActivity.this);
+								CategoryAdapter new_CateAdapter = new CategoryAdapter(
+										categories, PublishGoodsActivity.this);
 								category_Spinner.setAdapter(new_CateAdapter);
 							}
 						};
-						
+
 						new Thread(new Runnable() {
 							@Override
 							public void run() {
-								String json = HttpUtil.sendJson2Server(jsonObject,
-										URLs.GET_CATEGORY_URL);
+								String json = HttpUtil.sendJson2Server(
+										jsonObject, URLs.GET_CATEGORY_URL);
 								Message msg = Message.obtain();
 								msg.obj = json;
 								getCate_handler.sendMessage(msg);
@@ -649,26 +660,44 @@ public class PublishGoodsActivity extends Activity implements OnClickListener {
 		Dialog dialog = new AlertDialog.Builder(PublishGoodsActivity.this)
 				.setView(iv)
 				.setPositiveButton("删除", new DialogInterface.OnClickListener() {
+					@SuppressLint("ShowToast")
 					@Override
 					public void onClick(DialogInterface arg0, int arg1) {
 
 						// 1,从sd卡删除
 						File f1 = new File(list.get(index));// 缩略图
 						File f2 = new File(list.get(index).replace("_mini", ""));// 原图
-						if (f1.exists()) {
-							f1.delete();
-						}
-						if (f2.exists()) {
-							f2.delete();
-						}
 						// 2,向服务器发送请求...
-
-						// 3,从显示视图删除，并更新到gridview
-						list.remove(list.get(index));
-						myGirdView.setAdapter(new ImageAdapter(
-								PublishGoodsActivity.this, list,
-								metrics.widthPixels));
-
+						String param = "?imgid=" + JsonUtil.imgid;
+						JSONObject json = new JSONObject();
+						try {
+							json.put("imgid", JsonUtil.imgid);
+						} catch (JSONException e) {
+							throw new RuntimeException(e);
+						}
+						String jsonString = HttpUtil.sendJson2Server(json, URLs.DELIMG_URL);
+						JSONObject result = JsonUtil.parseJsonDelImg(jsonString);
+						int status = result.optInt("status");
+						String msg = "";
+						try {
+							msg = result.getString("msg");
+						} catch (JSONException e) {
+							e.printStackTrace();
+						}
+						if(status==1){
+							if (f1.exists()) {
+								f1.delete();
+							}
+							if (f2.exists()) {
+								f2.delete();
+							}
+							// 3,从显示视图删除，并更新到gridview
+							list.remove(list.get(index));
+							myGirdView.setAdapter(new ImageAdapter(
+									PublishGoodsActivity.this, list,
+									metrics.widthPixels));
+						}
+						Toast.makeText(PublishGoodsActivity.this, msg, Toast.LENGTH_SHORT);
 					}
 				})
 				.setNegativeButton("取消", new DialogInterface.OnClickListener() {
